requirements.txt
Flask==2.3.3
Flask-Cors==4.0.0
pycryptodome==3.20.0
PyNaCl==1.5.0
requests==2.32.3
pytest==8.3.3
